requirements.txt
nba_api>=1.4.1
requests>=2.28.0
python-telegram-bot>=20.0
schedule>=1.2.0
python-dotenv>=1.0.0
